warehouse/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.1